+++ packages/zhouhaifei-config/index.js
@@ -1,0 +1,7 @@
+const babel = require('./babel');
+const postcss = require('./postcss');
+
+module.exports = {
+  babel,
+  postcss,
+};