+++ postcss.config.js
@@ -1,0 +1,1 @@
+module.exports = require('./packages/zhouhaifei-config').postcss(0);